test-image-embedded-lunar: x86: add profiles
This adds two build profiles: `bootstrapped` and `default`.

The `bootstrapped` profile makes use of a prebuilt bootstrap archive
found in the given PPA. This allows quicker build times.

The `default` profile utilises debootstrap to build the image. This is
the default profile which is used when no profile is specified on the
command line.

LP: #2008216
Signed-off-by: Isaac True <isaac.true@canonical.com>