nemos-images-reference-lunar: preseed snaps in the development images
This change adds the checkbox and checkbox22 snaps to the reference
images built with the "development" profile. These snaps provide the
checkbox framework, which provides an automatic test suite inside the
resulting image. It also installs some additional tools which are needed
for the checkbox tests.

The snaps are installed using the preseeding functionality of snapd.
The snap files are downloaded during the Kiwi build process, so network
access to the Snap Store is required when building with the development
profile.

Due to the large size of the snap files, the partition sizes need to be
increased to accomodate the new tools and snaps.

Signed-off-by: Isaac True <isaac.true@canonical.com>