nemos-images-reference-lunar: qemu-{arm64,amd64}: add polkitd
This adds the polkitd utility, which allows adding policies which
determine what non-root users can do.

This also adds a default policy which allows the "admin" user to
power-off, reboot, and halt the system using dbus access to logind.

This can be demonstrated by calling the following command as the admin
user, which will shutdown the system:

    dbus-send --system --print-reply --dest=org.freedesktop.login1 \
        /org/freedesktop/login1 "org.freedesktop.login1.Manager.PowerOff" \
        boolean:true

Signed-off-by: Isaac True <isaac.true@canonical.com>