nemos-images-minimal-lunar: qemu-*: do not remove python3
Do not remove python3 packages from the rootfs, as these are required
for netplan.io.

This particular change actually has very little effect on the disk
usage; python3.10 doesn't exist in lunar (as it uses 3.11), so removing
the package was essentially a no-op before. Removing the python3 package
basically just removes the symlink for python3 to python3.11, which also
doesn't have an impact on the rootfs size.

Signed-off-by: Isaac True <isaac.true@canonical.com>