nemos-images-*-lunar: qemu-*: update grub parameters
Set the console to `serial` as the default `gfxterm` does not work in
the VM (as the VM does not have a graphics output).

Additionally, set the timeout to `0` to remove the boot delay. This
mostly invalidates the need for setting the console to `serial`, as the
console won't appear, but it's helpful to explicitly set these variables in
case a user wants to add a timeout for debugging purposes.

Signed-off-by: Isaac True <isaac.true@canonical.com>