ERL-270: github: workflows: nemos-images: build qemu-arm64 targets
This makes use of the kiwi box build plugin to create an arm64 VM.

Additionally, roll the spread step into the build step. This adds more
complication to the build step, but allows us to test each target
individually without running out of resources. Spread does not have a
way to limit how many VMs are created, and as more and more targets are
added, we quickly run out of resources on the GitHub runners.

Signed-off-by: Isaac True <isaac.true@canonical.com>