ERL-294: run_qemu.sh: use semi-persistent EFI variable storage
Instead of recreating the EFI variable storage with each invocation, the
default behaviour is to persist the EFI variable storage for a given VM
image in /tmp. Files in /tmp should be cleaned up by the system
automatically at reboot, so this shouldn't pollute the host.

This is useful for allowing pstore to make use of the EFI variable
storage for saving crash dumps.

Additionally, add an option for reverting to the previous behaviour of
always creating a temporary file for the EFI variables (-e), and another
option for overwriting any existing variable files (-o).

Signed-off-by: Isaac True <isaac.true@canonical.com>